fix: move cloneError to helpers.js
